requirements_abe.txt
attrs==25.3.0
cffi==2.0.0
Charm-Crypto @ file:///mnt/c/Users/makcay/Desktop/pesto-pacman/charm
cryptography==45.0.7
hypothesis==6.138.17
numpy==2.3.3
pillow==11.3.0
pycparser==2.23
pyparsing==2.4.0
setuptools==80.9.0
sortedcontainers==2.4.0
